Create Leaderboard Back-end logic
Two classes for time and average speed is created
to store sorted list of objects for each of them
in two different leaderboard.Also,a json class which
has methods to be used to serialize and deserialize
leaderboard objects.

Related issues: #4
