+++ .github/workflows/gameServer-build.yml
@@ -9,9 +9,9 @@ jobs:
 
     steps:
       - name: Clone Repo
-        uses: actions/checkout@v2
+        uses: actions/checkout@v3
       - name: Install JDK 11
-        uses: actions/setup-java@v2
+        uses: actions/setup-java@v3
         with:
           java-version: '11'
           distribution: 'temurin'
