+++ GameServer/src/main/java/main.java
@@ -1,6 +1,12 @@
+package GameServer.main.java;
+
 public class main {
 
     public static void main(String[] args) {
-        
+        /**to be implemented:
+         * create leaderboard object and serialize to a json file
+         * if a json file already exists then it is deserialized to a leaderboard object
+         * MQTT connection will be used to get time and average speed values,then create
+         * their correponding objects and add it to the leaderboard*/
     }
 }
